Cleanup location tracking logic into separate helper file.
